requirements.txt
fastapi
uvicorn
gradio
requests
pillow
transformers
llava
python-multipart